requirements.txt
streamlit
opencv-python
numpy
pillow
pandas
pymupdf
pytesseract
easyocr
